Update schoolVariantGen
Finished generation of tunnels that are not the initial one.

Still need to create a finder-thing to find what needs to be done.

The tunSessionDone is going to break it.

Not tested, however. Also, the generation for the random position for the initial tunnels are broken because I tried to modular divide by a negative.